portal-estado-am/requirements.txt
playwright==1.57.0
psycopg2-binary==2.9.9
